feat: useQuery 제외
