+++ src/apis/index.ts
@@ -1,3 +1,5 @@
 export { default as client } from "./client";
 export { default as nickNameCheck } from "./nickNameCheck";
 export { default as submitUserInformation } from "./submitUserInformation";
+export { default as getPlaceCategories} from "./useGetPlaceCategories";
+export { default as getTagCategories } from "./useGetTagCategories";
